+++ systest/backup/minio-large/docker-compose.yml
