+++ worker/backup_handler_test.go
@@ -10,7 +10,7 @@
  *     https://github.com/dgraph-io/dgraph/blob/master/licenses/DCL.txt
  */
 
-package backup
+package worker
 
 import (
 	"testing"
